Update to complete API yml and fix naming error
The former yml file was just a translation of the json schema. Now the
yml file describes the whole API. Information about pagination was
missing as well as HTTP error messages.

yml file had the wrong indention in arrays. Fixing this improved
readability.

There was a naming error calling the MOOChub schema "moochub_schema.json"
instead of "moochub-schema.json". This caused problems in matching the
correct files in the github repository.