Fix problem with quotation marks in yml
There was a problem that with generating the yml file by a
script the quotation marks were sometimes missing in the yml.
This caused a lot of trouble in reviewing changes in the
file, since all missing quotation marks were treated as
changes.

The script was updated and the file shows quotation marks now
where needed.