Improve readability and fix inconsistencies
There were some problems with data types of attributes.
Namely, some attributes can have a value of a defined type
or can be 'null'. In yml this was represented as an array.
As it is in the json. For consistency, the yml will now
only show the data type in the type attribute and has a
nullable attribute as before.

There were still issues with quotation marks that showed
changes where are none.

In the older version, empty attributes contained the value
'null'. This will be the case also for the newer version.
This way, no unnecessary tracking of changes occurs.