Fix further validation errors
There was a problem in the "Problem" part of the yml file.
the required attribute as well as readOnly and x-examples
were at the wrong level.

Fixing this erases validation errors in the yml file and
leads to fewer unwanted changes to be tracked.

Also corrects a mistake that tags now is an array as
intended.